Move sender threads to tokio
