+++ dds/src/dds/domain/dcps_service.rs
@@ -98,57 +98,6 @@ impl DcpsService {
         let quit = Arc::new(AtomicBool::new(false));
         let mut threads = Vec::new();
 
-        // //////////// Unicast metatraffic Communication send
-        {
-            let socket = UdpSocket::bind("0.0.0.0:0000").unwrap();
-
-            let mut metatraffic_unicast_transport_send = UdpTransportWrite::new(socket);
-            let task_quit = quit.clone();
-            let sedp_condvar_clone = sedp_condvar.clone();
-            threads.push(std::thread::spawn(move || loop {
-                if task_quit.load(atomic::Ordering::SeqCst) {
-                    break;
-                }
-                let _r = sedp_condvar_clone.wait_timeout(Duration::new(0, 500000000));
-                THE_DDS_DOMAIN_PARTICIPANT_FACTORY.get_participant_mut(
-                    &participant_guid_prefix,
-                    |dp| {
-                        if let Some(dp) = dp {
-                            send_user_defined_message(dp, &mut metatraffic_unicast_transport_send);
-                        }
-                    },
-                );
-            }));
-        }
-
-        // //////////// User-defined Communication send
-        {
-            let socket = UdpSocket::bind("0.0.0.0:0000").unwrap();
-            let mut default_unicast_transport_send = UdpTransportWrite::new(socket);
-            let task_quit = quit.clone();
-            let user_defined_data_send_condvar_clone = user_defined_data_send_condvar.clone();
-            threads.push(std::thread::spawn(move || loop {
-                if task_quit.load(atomic::Ordering::SeqCst) {
-                    break;
-                }
-
-                let _r = user_defined_data_send_condvar_clone
-                    .wait_timeout(Duration::new(0, 100_000_000));
-
-                THE_DDS_DOMAIN_PARTICIPANT_FACTORY.get_participant_mut(
-                    &participant_guid_prefix,
-                    |dp| {
-                        if let Some(dp) = dp {
-                            user_defined_communication_send(
-                                dp,
-                                &mut default_unicast_transport_send,
-                            );
-                        }
-                    },
-                )
-            }));
-        }
-
         let sedp_condvar = sedp_condvar.clone();
         let user_defined_data_send_condvar = user_defined_data_send_condvar.clone();
         Ok(DcpsService {
@@ -165,974 +114,5 @@ impl DcpsService {
 
     pub fn user_defined_data_send_condvar(&self) -> &DdsCondvar {
         &self.user_defined_data_send_condvar
-    }
-}
-
-fn remove_stale_writer_changes(writer: &mut DdsDataWriter<RtpsStatefulWriter>, now: Time) {
-    let timespan_duration = writer.get_qos().lifespan.duration;
-    writer.remove_change(|cc| DurationKind::Finite(now - cc.timestamp()) > timespan_duration);
-}
-
-fn send_message_best_effort_reader_locator(
-    reader_locator: &mut WriterAssociatedReaderLocator,
-    header: RtpsMessageHeader,
-    transport: &mut impl TransportWrite,
-    writer_id: EntityId,
-) {
-    let mut submessages = Vec::new();
-    while let Some(change) = reader_locator.next_unsent_change() {
-        // The post-condition:
-        // "( a_change BELONGS-TO the_reader_locator.unsent_changes() ) == FALSE"
-        // should be full-filled by next_unsent_change()
-        if let Some(cache_change) = change.cache_change() {
-            let info_ts_submessage = info_timestamp_submessage(cache_change.timestamp());
-            let data_submessage = cache_change.as_data_submessage(ENTITYID_UNKNOWN);
-            submessages.push(info_ts_submessage);
-            submessages.push(RtpsSubmessageKind::Data(data_submessage));
-        } else {
-            let gap_submessage = gap_submessage(writer_id, change.sequence_number());
-            submessages.push(gap_submessage);
-        }
-    }
-    if !submessages.is_empty() {
-        transport.write(
-            &RtpsMessage::new(header, submessages),
-            &[reader_locator.locator()],
-        )
-    }
-}
-
-fn send_message_best_effort_reader_proxy(
-    reader_proxy: &mut WriterAssociatedReaderProxy,
-    data_max_size_serialized: usize,
-    header: RtpsMessageHeader,
-    transport: &mut impl TransportWrite,
-) {
-    let info_dst = info_destination_submessage(reader_proxy.remote_reader_guid().prefix());
-    let mut submessages = vec![info_dst];
-
-    while !reader_proxy.unsent_changes().is_empty() {
-        // Note: The readerId is set to the remote reader ID as described in 8.4.9.2.12 Transition T12
-        // in confront to ENTITYID_UNKNOWN as described in 8.4.9.2.4 Transition T4
-        let reader_id = reader_proxy.remote_reader_guid().entity_id();
-        let change = reader_proxy.next_unsent_change();
-
-        if change.is_relevant() {
-            let cache_change = change.cache_change();
-            let timestamp = cache_change.timestamp();
-
-            if cache_change.data_value().len() > data_max_size_serialized {
-                let data_frag_submessage_list =
-                    cache_change.as_data_frag_submessages(data_max_size_serialized, reader_id);
-                for data_frag_submessage in data_frag_submessage_list {
-                    let info_dst =
-                        info_destination_submessage(reader_proxy.remote_reader_guid().prefix());
-
-                    let into_timestamp = info_timestamp_submessage(timestamp);
-                    let data_frag = RtpsSubmessageKind::DataFrag(data_frag_submessage);
-
-                    let submessages = vec![info_dst, into_timestamp, data_frag];
-
-                    transport.write(
-                        &RtpsMessage::new(header, submessages),
-                        reader_proxy.unicast_locator_list(),
-                    )
-                }
-            } else {
-                submessages.push(info_timestamp_submessage(timestamp));
-                submessages.push(RtpsSubmessageKind::Data(
-                    cache_change.as_data_submessage(reader_id),
-                ))
-            }
-        } else {
-            let gap_submessage: GapSubmessage = change
-                .cache_change()
-                .as_gap_message(reader_proxy.remote_reader_guid().entity_id());
-            submessages.push(RtpsSubmessageKind::Gap(gap_submessage));
-        }
-    }
-
-    // Send messages only if more than INFO_DST is added
-    if submessages.len() > 1 {
-        transport.write(
-            &RtpsMessage::new(header, submessages),
-            reader_proxy.unicast_locator_list(),
-        )
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn send_message_reliable_reader_proxy(
-    reader_proxy: &mut WriterAssociatedReaderProxy,
-    data_max_size_serialized: usize,
-    header: RtpsMessageHeader,
-    transport: &mut impl TransportWrite,
-    writer_id: EntityId,
-    first_sn: SequenceNumber,
-    last_sn: SequenceNumber,
-    heartbeat_period: Duration,
-) {
-    let reader_id = reader_proxy.remote_reader_guid().entity_id();
-
-    let info_dst = info_destination_submessage(reader_proxy.remote_reader_guid().prefix());
-
-    let mut submessages = vec![info_dst];
-
-    // Top part of the state machine - Figure 8.19 RTPS standard
-    if !reader_proxy.unsent_changes().is_empty() {
-        // Note: The readerId is set to the remote reader ID as described in 8.4.9.2.12 Transition T12
-        // in confront to ENTITYID_UNKNOWN as described in 8.4.9.2.4 Transition T4
-
-        while !reader_proxy.unsent_changes().is_empty() {
-            let change = reader_proxy.next_unsent_change();
-            // "a_change.status := UNDERWAY;" should be done by next_requested_change() as
-            // it's not done here to avoid the change being a mutable reference
-            // Also the post-condition:
-            // "( a_change BELONGS-TO the_reader_proxy.unsent_changes() ) == FALSE"
-            // should be full-filled by next_unsent_change()
-            if change.is_relevant() {
-                let cache_change = change.cache_change();
-                if cache_change.data_value().len() > data_max_size_serialized {
-                    directly_send_data_frag(
-                        reader_proxy,
-                        cache_change,
-                        writer_id,
-                        data_max_size_serialized,
-                        header,
-                        first_sn,
-                        last_sn,
-                        transport,
-                    );
-                    return;
-                } else {
-                    submessages.push(info_timestamp_submessage(cache_change.timestamp()));
-                    submessages.push(RtpsSubmessageKind::Data(
-                        cache_change.as_data_submessage(reader_id),
-                    ))
-                }
-            } else {
-                let gap_submessage: GapSubmessage = change.cache_change().as_gap_message(reader_id);
-
-                submessages.push(RtpsSubmessageKind::Gap(gap_submessage));
-            }
-        }
-
-        let heartbeat = reader_proxy
-            .heartbeat_machine()
-            .submessage(writer_id, first_sn, last_sn);
-        submessages.push(heartbeat);
-    } else if reader_proxy.unacked_changes().is_empty() {
-        // Idle
-    } else if reader_proxy
-        .heartbeat_machine()
-        .is_time_for_heartbeat(heartbeat_period)
-    {
-        let heartbeat = reader_proxy
-            .heartbeat_machine()
-            .submessage(writer_id, first_sn, last_sn);
-        submessages.push(heartbeat);
-    }
-
-    // Middle-part of the state-machine - Figure 8.19 RTPS standard
-    if !reader_proxy.requested_changes().is_empty() {
-        let reader_id = reader_proxy.remote_reader_guid().entity_id();
-
-        while !reader_proxy.requested_changes().is_empty() {
-            let change_for_reader = reader_proxy.next_requested_change();
-            // "a_change.status := UNDERWAY;" should be done by next_requested_change() as
-            // it's not done here to avoid the change being a mutable reference
-            // Also the post-condition:
-            // a_change BELONGS-TO the_reader_proxy.requested_changes() ) == FALSE
-            // should be full-filled by next_requested_change()
-            if change_for_reader.is_relevant() {
-                let cache_change = change_for_reader.cache_change();
-                if cache_change.data_value().len() > data_max_size_serialized {
-                    directly_send_data_frag(
-                        reader_proxy,
-                        cache_change,
-                        writer_id,
-                        data_max_size_serialized,
-                        header,
-                        first_sn,
-                        last_sn,
-                        transport,
-                    );
-                    return;
-                } else {
-                    submessages.push(info_timestamp_submessage(cache_change.timestamp()));
-                    submessages.push(RtpsSubmessageKind::Data(
-                        cache_change.as_data_submessage(reader_id),
-                    ))
-                }
-            } else {
-                let gap_submessage: GapSubmessage =
-                    change_for_reader.cache_change().as_gap_message(reader_id);
-
-                submessages.push(RtpsSubmessageKind::Gap(gap_submessage));
-            }
-        }
-        let heartbeat = reader_proxy
-            .heartbeat_machine()
-            .submessage(writer_id, first_sn, last_sn);
-        submessages.push(heartbeat);
-    }
-    // Send messages only if more or equal than INFO_DST and HEARTBEAT is added
-    if submessages.len() >= 2 {
-        transport.write(
-            &RtpsMessage::new(header, submessages),
-            reader_proxy.unicast_locator_list(),
-        )
-    }
-}
-
-fn gap_submessage<'a>(
-    writer_id: EntityId,
-    gap_sequence_number: SequenceNumber,
-) -> RtpsSubmessageKind<'a> {
-    RtpsSubmessageKind::Gap(GapSubmessage {
-        endianness_flag: true,
-        reader_id: ENTITYID_UNKNOWN,
-        writer_id,
-        gap_start: gap_sequence_number,
-        gap_list: SequenceNumberSet {
-            base: gap_sequence_number,
-            set: vec![],
-        },
-    })
-}
-
-fn info_timestamp_submessage<'a>(timestamp: Time) -> RtpsSubmessageKind<'a> {
-    RtpsSubmessageKind::InfoTimestamp(InfoTimestampSubmessage {
-        endianness_flag: true,
-        invalidate_flag: false,
-        timestamp: crate::implementation::rtps::messages::types::Time::new(
-            timestamp.sec(),
-            timestamp.nanosec(),
-        ),
-    })
-}
-
-fn info_destination_submessage<'a>(guid_prefix: GuidPrefix) -> RtpsSubmessageKind<'a> {
-    RtpsSubmessageKind::InfoDestination(InfoDestinationSubmessage {
-        endianness_flag: true,
-        guid_prefix,
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn directly_send_data_frag(
-    reader_proxy: &mut WriterAssociatedReaderProxy,
-    cache_change: &RtpsWriterCacheChange,
-    writer_id: EntityId,
-    data_max_size_serialized: usize,
-    header: RtpsMessageHeader,
-    first_sn: SequenceNumber,
-    last_sn: SequenceNumber,
-    transport: &mut impl TransportWrite,
-) {
-    let reader_id = reader_proxy.remote_reader_guid().entity_id();
-    let timestamp = cache_change.timestamp();
-
-    let mut data_frag_submessage_list = cache_change
-        .as_data_frag_submessages(data_max_size_serialized, reader_id)
-        .into_iter()
-        .peekable();
-
-    while let Some(data_frag_submessage) = data_frag_submessage_list.next() {
-        let writer_sn = data_frag_submessage.writer_sn;
-        let last_fragment_num = FragmentNumber::new(
-            u32::from(data_frag_submessage.fragment_starting_num)
-                + u16::from(data_frag_submessage.fragments_in_submessage) as u32
-                - 1,
-        );
-
-        let info_dst = info_destination_submessage(reader_proxy.remote_reader_guid().prefix());
-        let into_timestamp = info_timestamp_submessage(timestamp);
-        let data_frag = RtpsSubmessageKind::DataFrag(data_frag_submessage);
-
-        let is_last_fragment = data_frag_submessage_list.peek().is_none();
-        let submessages = if is_last_fragment {
-            let heartbeat_frag = reader_proxy.heartbeat_frag_machine().submessage(
-                writer_id,
-                writer_sn,
-                last_fragment_num,
-            );
-            vec![info_dst, into_timestamp, data_frag, heartbeat_frag]
-        } else {
-            let heartbeat = reader_proxy
-                .heartbeat_machine()
-                .submessage(writer_id, first_sn, last_sn);
-            vec![info_dst, into_timestamp, data_frag, heartbeat]
-        };
-        transport.write(
-            &RtpsMessage::new(header, submessages),
-            reader_proxy.unicast_locator_list(),
-        )
-    }
-}
-
-fn stateless_writer_send_message(
-    writer: &mut DdsDataWriter<RtpsStatelessWriter>,
-    header: RtpsMessageHeader,
-    transport: &mut impl TransportWrite,
-) {
-    let writer_id = writer.guid().entity_id();
-    for mut rl in &mut writer.reader_locator_list().into_iter() {
-        send_message_best_effort_reader_locator(&mut rl, header, transport, writer_id);
-    }
-}
-
-fn user_defined_stateful_writer_send_message(
-    writer: &mut DdsDataWriter<RtpsStatefulWriter>,
-    header: RtpsMessageHeader,
-    transport: &mut impl TransportWrite,
-) {
-    let data_max_size_serialized = writer.data_max_size_serialized();
-    let writer_id = writer.guid().entity_id();
-    let first_sn = writer
-        .change_list()
-        .iter()
-        .map(|x| x.sequence_number())
-        .min()
-        .unwrap_or_else(|| SequenceNumber::new(1));
-    let last_sn = writer
-        .change_list()
-        .iter()
-        .map(|x| x.sequence_number())
-        .max()
-        .unwrap_or_else(|| SequenceNumber::new(0));
-    let heartbeat_period = writer.heartbeat_period();
-    for reader_proxy in &mut writer.matched_reader_list() {
-        match reader_proxy.reliability() {
-            ReliabilityKind::BestEffort => send_message_best_effort_reader_proxy(
-                reader_proxy,
-                data_max_size_serialized,
-                header,
-                transport,
-            ),
-            ReliabilityKind::Reliable => send_message_reliable_reader_proxy(
-                reader_proxy,
-                data_max_size_serialized,
-                header,
-                transport,
-                writer_id,
-                first_sn,
-                last_sn,
-                heartbeat_period,
-            ),
-        }
-    }
-}
-
-fn discover_matched_writers(
-    domain_participant: &mut DdsDomainParticipant,
-    listener_sender: &tokio::sync::mpsc::Sender<ListenerTriggerKind>,
-) -> DdsResult<()> {
-    let samples = domain_participant
-        .get_builtin_subscriber_mut()
-        .stateful_data_reader_list_mut()
-        .iter_mut()
-        .find(|x| x.get_topic_name() == DCPS_PUBLICATION)
-        .unwrap()
-        .read::<DiscoveredWriterData>(
-            i32::MAX,
-            ANY_SAMPLE_STATE,
-            ANY_VIEW_STATE,
-            ANY_INSTANCE_STATE,
-            None,
-        )?;
-
-    for discovered_writer_data_sample in samples.into_iter() {
-        match discovered_writer_data_sample.sample_info.instance_state {
-            InstanceStateKind::Alive => {
-                if let Some(discovered_writer_data) = discovered_writer_data_sample.data {
-                    if !domain_participant.is_publication_ignored(
-                        discovered_writer_data
-                            .writer_proxy()
-                            .remote_writer_guid()
-                            .into(),
-                    ) {
-                        let remote_writer_guid_prefix = discovered_writer_data
-                            .writer_proxy()
-                            .remote_writer_guid()
-                            .prefix();
-                        let writer_parent_participant_guid =
-                            Guid::new(remote_writer_guid_prefix, ENTITYID_PARTICIPANT);
-
-                        if let Some((
-                            default_unicast_locator_list,
-                            default_multicast_locator_list,
-                        )) = domain_participant
-                            .discovered_participant_list()
-                            .into_iter()
-                            .find(|&(h, _)| {
-                                h == &InstanceHandle::from(writer_parent_participant_guid)
-                            })
-                            .map(|(_, discovered_participant_data)| {
-                                (
-                                    discovered_participant_data
-                                        .participant_proxy()
-                                        .default_unicast_locator_list()
-                                        .to_vec(),
-                                    discovered_participant_data
-                                        .participant_proxy()
-                                        .default_multicast_locator_list()
-                                        .to_vec(),
-                                )
-                            })
-                        {
-                            let domain_participant_guid = domain_participant.guid();
-                            for subscriber in domain_participant.user_defined_subscriber_list_mut()
-                            {
-                                subscriber_add_matched_writer(
-                                    subscriber,
-                                    &discovered_writer_data,
-                                    &default_unicast_locator_list,
-                                    &default_multicast_locator_list,
-                                    domain_participant_guid,
-                                    listener_sender,
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-            InstanceStateKind::NotAliveDisposed => {
-                let domain_participant_guid = domain_participant.guid();
-                for subscriber in domain_participant.user_defined_subscriber_list_mut() {
-                    let subscriber_guid = subscriber.guid();
-                    for data_reader in subscriber.stateful_data_reader_list_mut() {
-                        data_reader.remove_matched_writer(
-                            discovered_writer_data_sample.sample_info.instance_handle,
-                            domain_participant_guid,
-                            subscriber_guid,
-                            listener_sender,
-                        )
-                    }
-                }
-            }
-            InstanceStateKind::NotAliveNoWriters => todo!(),
-        }
-    }
-
-    Ok(())
-}
-
-pub fn subscriber_add_matched_writer(
-    user_defined_subscriber: &mut DdsSubscriber,
-    discovered_writer_data: &DiscoveredWriterData,
-    default_unicast_locator_list: &[Locator],
-    default_multicast_locator_list: &[Locator],
-    parent_participant_guid: Guid,
-    listener_sender: &tokio::sync::mpsc::Sender<ListenerTriggerKind>,
-) {
-    let is_discovered_writer_regex_matched_to_subscriber = if let Ok(d) = glob_to_regex(
-        &discovered_writer_data
-            .clone()
-            .dds_publication_data()
-            .partition()
-            .name,
-    ) {
-        d.is_match(&user_defined_subscriber.get_qos().partition.name)
-    } else {
-        false
-    };
-
-    let is_subscriber_regex_matched_to_discovered_writer =
-        if let Ok(d) = glob_to_regex(&user_defined_subscriber.get_qos().partition.name) {
-            d.is_match(
-                &discovered_writer_data
-                    .clone()
-                    .dds_publication_data()
-                    .partition()
-                    .name,
-            )
-        } else {
-            false
-        };
-
-    let is_partition_string_matched = discovered_writer_data
-        .clone()
-        .dds_publication_data()
-        .partition()
-        .name
-        == user_defined_subscriber.get_qos().partition.name;
-
-    if is_discovered_writer_regex_matched_to_subscriber
-        || is_subscriber_regex_matched_to_discovered_writer
-        || is_partition_string_matched
-    {
-        let user_defined_subscriber_qos = user_defined_subscriber.get_qos();
-        let user_defined_subscriber_guid = user_defined_subscriber.guid();
-        for data_reader in user_defined_subscriber.stateful_data_reader_list_mut() {
-            data_reader.add_matched_writer(
-                discovered_writer_data,
-                default_unicast_locator_list,
-                default_multicast_locator_list,
-                &user_defined_subscriber_qos,
-                user_defined_subscriber_guid,
-                parent_participant_guid,
-                listener_sender,
-            )
-        }
-    }
-}
-
-fn discover_matched_participants(
-    domain_participant: &mut DdsDomainParticipant,
-    sedp_condvar: &DdsCondvar,
-) -> DdsResult<()> {
-    while let Ok(samples) = domain_participant
-        .get_builtin_subscriber_mut()
-        .stateless_data_reader_list_mut()
-        .iter_mut()
-        .find(|x| x.get_topic_name() == DCPS_PARTICIPANT)
-        .unwrap()
-        .read(
-            1,
-            &[SampleStateKind::NotRead],
-            ANY_VIEW_STATE,
-            ANY_INSTANCE_STATE,
-            None,
-        )
-    {
-        for discovered_participant_data_sample in samples.into_iter() {
-            if let Some(discovered_participant_data) = discovered_participant_data_sample.data {
-                add_discovered_participant(domain_participant, discovered_participant_data);
-                sedp_condvar.notify_all();
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn add_discovered_participant(
-    domain_participant: &mut DdsDomainParticipant,
-    discovered_participant_data: SpdpDiscoveredParticipantData,
-) {
-    if ParticipantDiscovery::new(
-        &discovered_participant_data,
-        domain_participant.get_domain_id(),
-        domain_participant.get_domain_tag(),
-    )
-    .is_ok()
-        && !domain_participant
-            .is_participant_ignored(discovered_participant_data.get_serialized_key().into())
-    {
-        add_matched_publications_detector(
-            domain_participant
-                .get_builtin_publisher_mut()
-                .stateful_data_writer_list_mut()
-                .iter_mut()
-                .find(|x| x.get_topic_name() == DCPS_PUBLICATION)
-                .unwrap(),
-            &discovered_participant_data,
-        );
-
-        add_matched_publications_announcer(
-            domain_participant
-                .get_builtin_subscriber_mut()
-                .stateful_data_reader_list_mut()
-                .iter_mut()
-                .find(|x| x.get_topic_name() == DCPS_PUBLICATION)
-                .unwrap(),
-            &discovered_participant_data,
-        );
-
-        add_matched_subscriptions_detector(
-            domain_participant
-                .get_builtin_publisher_mut()
-                .stateful_data_writer_list_mut()
-                .iter_mut()
-                .find(|x| x.get_topic_name() == DCPS_SUBSCRIPTION)
-                .unwrap(),
-            &discovered_participant_data,
-        );
-
-        add_matched_subscriptions_announcer(
-            domain_participant
-                .get_builtin_subscriber_mut()
-                .stateful_data_reader_list_mut()
-                .iter_mut()
-                .find(|x| x.get_topic_name() == DCPS_SUBSCRIPTION)
-                .unwrap(),
-            &discovered_participant_data,
-        );
-
-        add_matched_topics_detector(
-            domain_participant
-                .get_builtin_publisher_mut()
-                .stateful_data_writer_list_mut()
-                .iter_mut()
-                .find(|x| x.get_topic_name() == DCPS_TOPIC)
-                .unwrap(),
-            &discovered_participant_data,
-        );
-
-        add_matched_topics_announcer(
-            domain_participant
-                .get_builtin_subscriber_mut()
-                .stateful_data_reader_list_mut()
-                .iter_mut()
-                .find(|x| x.get_topic_name() == DCPS_TOPIC)
-                .unwrap(),
-            &discovered_participant_data,
-        );
-
-        domain_participant.discovered_participant_add(
-            discovered_participant_data.get_serialized_key().into(),
-            discovered_participant_data,
-        );
-    }
-}
-
-fn add_matched_subscriptions_announcer(
-    reader: &mut DdsDataReader<RtpsStatefulReader>,
-    discovered_participant_data: &SpdpDiscoveredParticipantData,
-) {
-    if discovered_participant_data
-        .participant_proxy()
-        .available_builtin_endpoints()
-        .has(BuiltinEndpointSet::BUILTIN_ENDPOINT_SUBSCRIPTIONS_ANNOUNCER)
-    {
-        let remote_writer_guid = Guid::new(
-            discovered_participant_data
-                .participant_proxy()
-                .guid_prefix(),
-            ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_ANNOUNCER,
-        );
-        let remote_group_entity_id = ENTITYID_UNKNOWN;
-        let data_max_size_serialized = None;
-
-        let proxy = RtpsWriterProxy::new(
-            remote_writer_guid,
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_unicast_locator_list(),
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_multicast_locator_list(),
-            data_max_size_serialized,
-            remote_group_entity_id,
-        );
-        reader.matched_writer_add(proxy);
-    }
-}
-
-fn add_matched_subscriptions_detector(
-    writer: &mut DdsDataWriter<RtpsStatefulWriter>,
-    discovered_participant_data: &SpdpDiscoveredParticipantData,
-) {
-    if discovered_participant_data
-        .participant_proxy()
-        .available_builtin_endpoints()
-        .has(BuiltinEndpointSet::BUILTIN_ENDPOINT_SUBSCRIPTIONS_DETECTOR)
-    {
-        let remote_reader_guid = Guid::new(
-            discovered_participant_data
-                .participant_proxy()
-                .guid_prefix(),
-            ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_DETECTOR,
-        );
-        let remote_group_entity_id = ENTITYID_UNKNOWN;
-        let expects_inline_qos = false;
-        let proxy = RtpsReaderProxy::new(
-            remote_reader_guid,
-            remote_group_entity_id,
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_unicast_locator_list(),
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_multicast_locator_list(),
-            expects_inline_qos,
-            true,
-            ReliabilityKind::Reliable,
-            DurabilityKind::TransientLocal,
-        );
-        writer.matched_reader_add(proxy);
-    }
-}
-
-fn add_matched_publications_announcer(
-    reader: &mut DdsDataReader<RtpsStatefulReader>,
-    discovered_participant_data: &SpdpDiscoveredParticipantData,
-) {
-    if discovered_participant_data
-        .participant_proxy()
-        .available_builtin_endpoints()
-        .has(BuiltinEndpointSet::BUILTIN_ENDPOINT_PUBLICATIONS_ANNOUNCER)
-    {
-        let remote_writer_guid = Guid::new(
-            discovered_participant_data
-                .participant_proxy()
-                .guid_prefix(),
-            ENTITYID_SEDP_BUILTIN_PUBLICATIONS_ANNOUNCER,
-        );
-        let remote_group_entity_id = ENTITYID_UNKNOWN;
-        let data_max_size_serialized = None;
-
-        let proxy = RtpsWriterProxy::new(
-            remote_writer_guid,
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_unicast_locator_list(),
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_multicast_locator_list(),
-            data_max_size_serialized,
-            remote_group_entity_id,
-        );
-
-        reader.matched_writer_add(proxy);
-    }
-}
-
-fn add_matched_publications_detector(
-    writer: &mut DdsDataWriter<RtpsStatefulWriter>,
-    discovered_participant_data: &SpdpDiscoveredParticipantData,
-) {
-    if discovered_participant_data
-        .participant_proxy()
-        .available_builtin_endpoints()
-        .has(BuiltinEndpointSet::BUILTIN_ENDPOINT_PUBLICATIONS_DETECTOR)
-    {
-        let remote_reader_guid = Guid::new(
-            discovered_participant_data
-                .participant_proxy()
-                .guid_prefix(),
-            ENTITYID_SEDP_BUILTIN_PUBLICATIONS_DETECTOR,
-        );
-        let remote_group_entity_id = ENTITYID_UNKNOWN;
-        let expects_inline_qos = false;
-        let proxy = RtpsReaderProxy::new(
-            remote_reader_guid,
-            remote_group_entity_id,
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_unicast_locator_list(),
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_multicast_locator_list(),
-            expects_inline_qos,
-            true,
-            ReliabilityKind::Reliable,
-            DurabilityKind::TransientLocal,
-        );
-        writer.matched_reader_add(proxy);
-    }
-}
-
-fn add_matched_topics_announcer(
-    reader: &mut DdsDataReader<RtpsStatefulReader>,
-    discovered_participant_data: &SpdpDiscoveredParticipantData,
-) {
-    if discovered_participant_data
-        .participant_proxy()
-        .available_builtin_endpoints()
-        .has(BuiltinEndpointSet::BUILTIN_ENDPOINT_TOPICS_ANNOUNCER)
-    {
-        let remote_writer_guid = Guid::new(
-            discovered_participant_data
-                .participant_proxy()
-                .guid_prefix(),
-            ENTITYID_SEDP_BUILTIN_TOPICS_ANNOUNCER,
-        );
-        let remote_group_entity_id = ENTITYID_UNKNOWN;
-        let data_max_size_serialized = None;
-
-        let proxy = RtpsWriterProxy::new(
-            remote_writer_guid,
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_unicast_locator_list(),
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_multicast_locator_list(),
-            data_max_size_serialized,
-            remote_group_entity_id,
-        );
-        reader.matched_writer_add(proxy);
-    }
-}
-
-fn add_matched_topics_detector(
-    writer: &mut DdsDataWriter<RtpsStatefulWriter>,
-    discovered_participant_data: &SpdpDiscoveredParticipantData,
-) {
-    if discovered_participant_data
-        .participant_proxy()
-        .available_builtin_endpoints()
-        .has(BuiltinEndpointSet::BUILTIN_ENDPOINT_TOPICS_DETECTOR)
-    {
-        let remote_reader_guid = Guid::new(
-            discovered_participant_data
-                .participant_proxy()
-                .guid_prefix(),
-            ENTITYID_SEDP_BUILTIN_TOPICS_DETECTOR,
-        );
-        let remote_group_entity_id = ENTITYID_UNKNOWN;
-        let expects_inline_qos = false;
-        let proxy = RtpsReaderProxy::new(
-            remote_reader_guid,
-            remote_group_entity_id,
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_unicast_locator_list(),
-            discovered_participant_data
-                .participant_proxy()
-                .metatraffic_multicast_locator_list(),
-            expects_inline_qos,
-            true,
-            ReliabilityKind::Reliable,
-            DurabilityKind::TransientLocal,
-        );
-        writer.matched_reader_add(proxy);
-    }
-}
-
-pub fn receive_builtin_message(
-    domain_participant: &mut DdsDomainParticipant,
-    message: RtpsMessage,
-    locator: Locator,
-    sedp_condvar: &DdsCondvar,
-    listener_sender: &tokio::sync::mpsc::Sender<ListenerTriggerKind>,
-) {
-    domain_participant
-        .receive_builtin_data(locator, message, listener_sender)
-        .ok();
-
-    discover_matched_participants(domain_participant, sedp_condvar).ok();
-    domain_participant
-        .discover_matched_readers(listener_sender)
-        .ok();
-    discover_matched_writers(domain_participant, listener_sender).ok();
-    domain_participant
-        .discover_matched_topics(listener_sender)
-        .ok();
-}
-
-fn send_user_defined_message(
-    domain_participant: &mut DdsDomainParticipant,
-    metatraffic_unicast_transport_send: &mut impl TransportWrite,
-) {
-    let header = RtpsMessageHeader {
-        protocol: ProtocolId::PROTOCOL_RTPS,
-        version: domain_participant.protocol_version(),
-        vendor_id: domain_participant.vendor_id(),
-        guid_prefix: domain_participant.guid().prefix(),
-    };
-
-    let _now = domain_participant.get_current_time();
-    stateless_writer_send_message(
-        domain_participant
-            .get_builtin_publisher_mut()
-            .stateless_data_writer_list_mut()
-            .iter_mut()
-            .find(|x| x.get_type_name() == SpdpDiscoveredParticipantData::type_name())
-            .unwrap(),
-        header,
-        metatraffic_unicast_transport_send,
-    );
-
-    user_defined_stateful_writer_send_message(
-        domain_participant
-            .get_builtin_publisher_mut()
-            .stateful_data_writer_list_mut()
-            .iter_mut()
-            .find(|x| x.get_type_name() == DiscoveredWriterData::type_name())
-            .unwrap(),
-        header,
-        metatraffic_unicast_transport_send,
-    );
-
-    user_defined_stateful_writer_send_message(
-        domain_participant
-            .get_builtin_publisher_mut()
-            .stateful_data_writer_list_mut()
-            .iter_mut()
-            .find(|x| x.get_type_name() == DiscoveredReaderData::type_name())
-            .unwrap(),
-        header,
-        metatraffic_unicast_transport_send,
-    );
-
-    user_defined_stateful_writer_send_message(
-        domain_participant
-            .get_builtin_publisher_mut()
-            .stateful_data_writer_list_mut()
-            .iter_mut()
-            .find(|x| x.get_type_name() == DiscoveredTopicData::type_name())
-            .unwrap(),
-        header,
-        metatraffic_unicast_transport_send,
-    );
-
-    for stateful_readers in domain_participant
-        .get_builtin_subscriber_mut()
-        .stateful_data_reader_list_mut()
-    {
-        stateful_readers.send_message(header, metatraffic_unicast_transport_send)
-    }
-}
-
-fn user_defined_communication_send(
-    domain_participant: &mut DdsDomainParticipant,
-    default_unicast_transport_send: &mut impl TransportWrite,
-) {
-    let header = RtpsMessageHeader {
-        protocol: ProtocolId::PROTOCOL_RTPS,
-        version: domain_participant.protocol_version(),
-        vendor_id: domain_participant.vendor_id(),
-        guid_prefix: domain_participant.guid().prefix(),
-    };
-    let now = domain_participant.get_current_time();
-
-    for publisher in domain_participant.user_defined_publisher_list_mut() {
-        for data_writer in publisher.stateful_data_writer_list_mut() {
-            let writer_id = data_writer.guid().entity_id();
-            let data_max_size_serialized = data_writer.data_max_size_serialized();
-            let heartbeat_period = data_writer.heartbeat_period();
-            let first_sn = data_writer
-                .change_list()
-                .iter()
-                .map(|x| x.sequence_number())
-                .min()
-                .unwrap_or(SequenceNumber::new(1));
-            let last_sn = data_writer
-                .change_list()
-                .iter()
-                .map(|x| x.sequence_number())
-                .max()
-                .unwrap_or_else(|| SequenceNumber::new(0));
-            remove_stale_writer_changes(data_writer, now);
-            for reader_proxy in &mut data_writer.matched_reader_list() {
-                match reader_proxy.reliability() {
-                    ReliabilityKind::BestEffort => send_message_best_effort_reader_proxy(
-                        reader_proxy,
-                        data_max_size_serialized,
-                        header,
-                        default_unicast_transport_send,
-                    ),
-                    ReliabilityKind::Reliable => send_message_reliable_reader_proxy(
-                        reader_proxy,
-                        data_max_size_serialized,
-                        header,
-                        default_unicast_transport_send,
-                        writer_id,
-                        first_sn,
-                        last_sn,
-                        heartbeat_period,
-                    ),
-                }
-            }
-        }
-    }
-
-    for subscriber in domain_participant.user_defined_subscriber_list_mut() {
-        for data_reader in subscriber.stateful_data_reader_list_mut() {
-            data_reader.send_message(header, default_unicast_transport_send)
-        }
     }
 }
